requirements.txt
# Core utilities
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0

# Pinecone vector database
pinecone>=4.0.0

# CLIP and transformers
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Google Generative AI
google-genai>=0.1.0

# Web UI
flask>=3.0.0

# Optional utilities
tqdm>=4.65.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0
setuptools~=80.9.0